requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
Werkzeug==2.3.7
Pillow # 
python-dotenv # 
flask-login
openpyxl
uwsgi